Selection Buttons/ Selection Panel now functional
When a selection button [Devices, Test Cases, Saved Results] is pressed...the copy (text) of the button is used to determine the name of the database that needs to be accessed. The name is then forwarded to main.py on the Flask Server from where it is then forwarded to a python script containing various database functions. The name is used to connect to the correct server and then trimmed to identify both the table and the SELECT criteria. The database is queried and the results are ultimately turned into a list of item names. That list is sent back to main from where it is sent back to the SelectionPanelController.js file. The list is iterated through and for each item in the list, a new list element is created and appended to the Selection Panel. Every time the a selection button is click the Right Panel is emptied of all child elements (ie. the list is cleared).